LayoutTests/fast/css-grid-layout/grid-disabled-by-default-expected.txt

Verifies that display: -webkit-grid does nothing by default.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementById('grid').style.display is ''
PASS successfullyParsed is true

TEST COMPLETE
